requirements.txt
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
pyproj==3.6.1
matplotlib==3.9.2
pytesseract==0.3.13
PyYAML==6.0.1
scipy==1.13.1
